challengeDjango/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.4
filelock==3.13.3
pbr==6.0.0
platformdirs==4.2.0
setuptools==69.2.0
sqlparse==0.5.0
stevedore==5.2.0
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wheel==0.43.0
djangorestframework
djangorestframework-simplejwt
requests
PyJWT
